+++ src/main/java/dk/sdu/se_f22/ProductModule/Infrastructure/data/TokenParameterStore.java
@@ -1,0 +1,12 @@
+package dk.sdu.se_f22.ProductModule.Infrastructure.data;
+
+class TokenParameterStore {
+
+	public static void saveTokenParameter(TokenParameter tp) {
+
+	}
+
+	public static TokenParameter loadTokenParameter() {
+		return new TokenParameter(" ", "('./?!')");
+	}
+}
