+++ src/main/java/dk/sdu/se_f22/ProductModule/Infrastructure/domain/ProductInfSearch.java
@@ -1,8 +1,5 @@
-package dk.sdu.se_f22.pim2.domain;
+package dk.sdu.se_f22.ProductModule.Infrastructure.domain;
 
-import dk.sdu.se_f22.pim2.Product;
-
-import java.util.ArrayList;
 import java.util.List;
 
 public interface ProductInfSearch {
