+++ src/main/java/dk/sdu/se_f22/ProductModule/Infrastructure/domain/ProductInfSearchImpl.java
@@ -1,0 +1,12 @@
+package dk.sdu.se_f22.ProductModule.Infrastructure.domain;
+
+import java.util.List;
+
+public class ProductInfSearchImpl implements ProductInfSearch {
+
+    @Override
+    public List<Product> searchProducts(List<String> tokens) {
+        // return PIM3.searchProducts(tokens);
+        return null;
+    }
+}
