+++ src/main/java/dk/sdu/se_f22/SharedLibrary/models/Product.java
@@ -1,0 +1,4 @@
+package dk.sdu.se_f22.SharedLibrary.models;
+
+public class Product {
+}
