--- src/main/java/dk/sdu/se_f22/pim2/Product.java
@@ -1,5 +1,0 @@
-package dk.sdu.se_f22.pim2;
-
-public class Product {
-	public String description = "My awsome Product is boring!";
-}
