--- src/main/java/dk/sdu/se_f22/pim2/data/TokenParameterStore.java
@@ -1,12 +1,0 @@
-package dk.sdu.se_f22.pim2.data;
-
-class TokenParameterStore {
-
-	public static void saveTokenParameter(TokenParameter tp) {
-
-	}
-
-	public static TokenParameter loadTokenParameter() {
-		return new TokenParameter(" ", "('./?!')");
-	}
-}
