--- src/main/java/dk/sdu/se_f22/pim2/domain/ProductInfIndex.java
@@ -1,16 +1,0 @@
-package dk.sdu.se_f22.pim2.domain;
-
-import dk.sdu.se_f22.pim2.Product;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public interface ProductInfIndex {
-
-void indexProducts(List<Product> products);
-
-
-
-
-
-}
