--- src/main/java/dk/sdu/se_f22/pim2/domain/ProductInfSearchImpl.java
@@ -1,14 +1,0 @@
-package dk.sdu.se_f22.pim2.domain;
-
-import dk.sdu.se_f22.pim2.Product;
-
-import java.util.List;
-
-public class ProductInfSearchImpl implements ProductInfSearch {
-
-    @Override
-    public List<Product> searchProducts(List<String> tokens) {
-        // return PIM3.searchProducts(tokens);
-        return null;
-    }
-}
